Move AssetDispatcher tests to correct test file.
